Create simple matrix to represent the game field
Created a SimpleMatrix class which wraps around a one-dimensional
std::vector. This allows us to separate concerns as we will be creating
different types of game fields and also saves us the hassle of having to
deal with nested vectors. SimpleMatrix is a class template as we are going
to use different data types to store our cell map in the future.